+++ aspecd/plotting.py
@@ -1568,6 +1568,8 @@ class MultiPlotter(Plotter):
         super().plot()
         self._set_axes_labels()
         self.properties.apply(plotter=self)
+        # Update/redraw legend after having set properties
+        self._set_legend()
 
     def _check_for_applicability(self):
         if not self.datasets:
@@ -2730,12 +2732,23 @@ class MultiPlot1DProperties(MultiPlotProperties):
         For the properties that can be set this way, see the documentation
         of the :class:`aspecd.plotting.LineProperties` class.
 
+    colormap : :class:`str`
+        NAme of the colormap to use for colouring the individual drawings
+
     Raises
     ------
     aspecd.exceptions.MissingPlotterError
         Raised if no plotter is provided.
 
+
+    .. versionchanged:: 0.8
+        Added attribute :attr:`colormap`
+
     """
+
+    def __init__(self):
+        super().__init__()
+        self.colormap = None
 
     def add_drawing(self):
         """
@@ -2759,6 +2772,33 @@ class MultiPlot1DProperties(MultiPlotProperties):
             rc_property = 'lines.' + key
             if rc_property in mpl.rcParams.keys():
                 setattr(drawing_properties, key, mpl.rcParams[rc_property])
+
+    def apply(self, plotter=None):
+        """
+        Apply properties to plot.
+
+        The main difference to the parent class: if you set a colormap
+        property, the lines will be coloured according to the colormap.
+        Note that this needs to be done on this level, as we need to know
+        how many drawings (i.e. lines) there are to colour.
+
+        Parameters
+        ----------
+        plotter: :class:`aspecd.plotting.MultiPlotter`
+            Plotter the properties should be applied to.
+
+        Raises
+        ------
+        aspecd.exceptions.MissingPlotterError
+            Raised if no plotter is provided.
+
+        """
+        super().apply(plotter=plotter)
+        if hasattr(plotter, 'drawings') and self.colormap:
+            idx = len(self.drawings)
+            colors = plt.get_cmap(self.colormap, idx)
+            for idx, drawing in enumerate(plotter.drawings):
+                self.drawings[idx].color = colors(idx)
 
 
 class CompositePlotProperties(PlotProperties):
@@ -3102,6 +3142,11 @@ class LegendProperties(aspecd.utils.Properties):
 
         Default: ``plt.rcParams['font.size']``
 
+    ncol : :class:`int`
+        Number of columns of the legend
+
+        Default: 1
+
     Raises
     ------
     aspecd.exceptions.MissingLegendError
@@ -3111,6 +3156,9 @@ class LegendProperties(aspecd.utils.Properties):
     .. versionchanged:: 0.7
         Added attributes :attr:`labelspacing` and :attr:`fontsize`
 
+    .. versionchanged:: 0.8
+        Added attribute :attr:`ncol`
+
     """
 
     def __init__(self):
@@ -3119,6 +3167,7 @@ class LegendProperties(aspecd.utils.Properties):
         self.frameon = True
         self.labelspacing = 0.5
         self.fontsize = plt.rcParams['font.size']
+        self.ncol = 1
         self._exclude = ['location']
         self._exclude_from_to_dict = ['location']
 
@@ -3138,12 +3187,12 @@ class LegendProperties(aspecd.utils.Properties):
         Parameters
         ----------
         legend: :class:`matplotlib.legend.Legend`
-            Plotter the properties should be applied to.
+            Legend the properties should be applied to.
 
         Raises
         ------
-        aspecd.exceptions.MissingFigureError
-            Raised if no figure is provided.
+        aspecd.exceptions.MissingLegendError
+            Raised if no legend is provided.
 
         """
         if not legend: